Edit Role and Applay Role in Api File

+++ routes/custom-route/api-dashboard-auth.php
@@ -24,43 +24,44 @@ Route::middleware(['auth:sanctum', 'type.admin'])->group(function () {
 
 
     # # # # # # # # # # # # # # # Home # # # # # # # # # # # # # # # 
-    Route::get('/home', [HomeDashApiController::class, 'index']);
+    Route::get('/home', [HomeDashApiController::class, 'index'])->middleware('check.role:HomeChart');
     # # # # # # # # # # # # # # # End Home # # # # # # # # # # # # # # # 
 
 
     # # # # # # # # # # # # # # # Message # # # # # # # # # # # # # # # 
     Route::group(['prefix' => 'message'], function () {
-        Route::get('/', [MessageDashApiController::class, 'index']);
-        Route::delete('/{message}', [MessageDashApiController::class, 'delete']);
-        Route::put('/{message}', [MessageDashApiController::class, 'edit']);
-        Route::get('/{message}', [MessageDashApiController::class, 'show']);
+        Route::get('/', [MessageDashApiController::class, 'index'])->middleware('check.role:ReadMessage');
+        Route::delete('/{message}', [MessageDashApiController::class, 'delete'])->middleware('check.role:DeleteMessage');
+        Route::put('/{message}', [MessageDashApiController::class, 'edit'])->middleware('check.role:EditMessage');
+        Route::get('/{message}', [MessageDashApiController::class, 'show'])->middleware('check.role:ReadMessage');
     });
     # # # # # # # # # # # # # # # End Message # # # # # # # # # # # # # # # 
 
 
     # # # # # # # # # # # # # # # Admin # # # # # # # # # # # # # # # 
     Route::group(['prefix' => 'admin'], function () {
-        Route::get('/', [AdminDashApiController::class, 'index']);
-        Route::get('/role', [AdminDashApiController::class, 'role']);
-        Route::get('/{admin}', [AdminDashApiController::class, 'show']);
-        Route::post('/', [AdminDashApiController::class, 'create']);
-        Route::put('/{admin}/active', [AdminDashApiController::class, 'active']);
-        Route::put('/{admin}/disActive', [AdminDashApiController::class, 'disActive']);
-        Route::put('/{admin}/banned', [AdminDashApiController::class, 'banned']);
-        Route::put('/{admin}/reset', [AdminDashApiController::class, 'resetPassword']);
+        Route::get('/', [AdminDashApiController::class, 'index'])->middleware('check.role:ReadAdmin');
+        Route::get('/role', [AdminDashApiController::class, 'role'])->middleware('check.role:CreateAdmin');
+        Route::get('/{admin}', [AdminDashApiController::class, 'show'])->middleware('check.role:ReadAdmin');
+        Route::post('/', [AdminDashApiController::class, 'create'])->middleware('check.role:CreateAdmin');
+        Route::put('/{admin}/active', [AdminDashApiController::class, 'active'])->middleware('check.role:ActiveAdmin');
+        Route::put('/{admin}/disActive', [AdminDashApiController::class, 'disActive'])->middleware('check.role:DisActiveAdmin');
+        Route::put('/{admin}/banned', [AdminDashApiController::class, 'banned'])->middleware('check.role:BannedAdmin');
+        Route::put('/{admin}/reset', [AdminDashApiController::class, 'resetPassword'])->middleware('check.role:ResetPasswordAdmin');
+        Route::put('/{admin}/role', [AdminDashApiController::class, 'changeAdminRole'])->middleware('check.role:EditRoleAdmin');
     });
     # # # # # # # # # # # # # # # End Admin # # # # # # # # # # # # # # # 
 
 
 
     # # # # # # # # # # # # # # # Roles # # # # # # # # # # # # # # # 
-    Route::group(['prefix' => 'role','middleware'=>'check.role:ReadMessage'], function () {
-        Route::get('/', [RoleController::class, 'index']);
-        Route::post('/', [RoleController::class, 'create']);
-        Route::delete('/{role}', [RoleController::class, 'delete']);
-        Route::PUT('/{role}', [RoleController::class, 'edit']);
-        Route::get('/permissions', [RoleController::class, 'permissions']);
-        Route::get('/{role}', [RoleController::class, 'show']);
+    Route::group(['prefix' => 'role'], function () {
+        Route::get('/', [RoleController::class, 'index'])->middleware('check.role:ReadRole');
+        Route::post('/', [RoleController::class, 'create'])->middleware('check.role:CreateRole');
+        Route::delete('/{role}', [RoleController::class, 'delete'])->middleware('check.role:DeleteRole');
+        Route::PUT('/{role}', [RoleController::class, 'edit'])->middleware('check.role:EditRole');
+        Route::get('/permissions', [RoleController::class, 'permissions'])->middleware('check.role:CreateRole');
+        Route::get('/{role}', [RoleController::class, 'show'])->middleware('check.role:ReadRole');
     });
     # # # # # # # # # # # # # # # End Roles # # # # # # # # # # # # # # #
